uint8_t bug

+++ dmludp/include/dmludp.h
@@ -103,8 +103,8 @@ void dmludp_config_free(dmludp_config *config);
 ////////////////////////////////////////
 // Extracts version, type, source / destination connection ID and address
 // verification token from the packet in |buf|.
-int dmludp_header_info(const uint8_t *buf, size_t buf_len, uint8_t *type, int *pkt_num);
-
+// int dmludp_header_info(const uint8_t *buf, size_t buf_len, uint8_t *type, int *pkt_num);
+int dmludp_header_info(const uint8_t *buf, size_t buf_len, int *type, int *pkt_num);
 
 typedef struct dmludp_conn dmludp_conn;
